SettingsButton: refactor to use sx prop instead of makeStyles
