--- src/components/annotator/SettingsButton/SettingsButton/index.ts
@@ -1,1 +1,0 @@
-export { SettingsButton } from "./SettingsButton";
