+++ src/components/annotator/SettingsDialog/index.ts
